Atlas - dockerization
This is not an absolute necessity to dockerize this new webapp, but I'm
requiring it early for runtime testing in a whole env (mfapp + client
part of the atlas).

Note: it will also require a PR on the datadir, because one won't be
able to generate the docker image without a datadir.