Harmonize spring and other important dependency versions
Harmonize spring and other important dependency versions
across geoOrchestra modules (not on submodules i.e. geoserver
and geonetwork).

Depending on a released version of spring-boot-dependencies provides
a curated list of dependency versions for all modules.

See https://github.com/spring-projects/spring-boot/blob/v1.5.19.RELEASE/spring-boot-dependencies/pom.xml,
which is the last release of the 1.x series, defining dependencies for
Spring 4.3.22.RELEASE and related dependencies that are known to
work well with that spring version.

This is common practice when working on spring-boot projects.

Other dependency versions harmonized are:

- Spring-security: 4.2.11.RELEASE, by means of the parent pom.
- Postgresql: 42.2.5.
- GeoTools: 19.4: this is the latest gt release that uses com.vividsolutions:jts.
  Upgrading to 20+ implies porting mapfishapp to use org.locationtech.jts:jts-core.

All additional dependencies (those not listed in the
spring-boot-dependeices pom) are added to the dependencyManagement
section in the root pom, and the individual module dependency
declarations have the version removed, so that a single version of each
dependency is ensured across all modules.

A special case is extractorapp, which overrides the gt.version property
declared in the root pom to stay at version 9.2, as it uses the gt-mif
module that seems to no longer exist, and hence require additional work
to upgrade to a more recent geotools version.

Note: comments were removed from
/mapfishapp/ws/addons/annotation/config.json and
/mapfishapp/ws/addons/magnifier/config.json. The JSON
spec does not allow comments in JSON and if we let them
and strip them out would be hurting interoperability.