atlas - restoring compatible dependencies (#3167)
See #3167 for the motivation.

Tests: compile / runtime in a minimal docker compo.

Note: see comments in the issue, I first tried to upgrade the libraries
to more recent versions, but this lead to other issues that have to be
addressed more thoroughly, and we are in a hurry with the release right
now.